Added logic to choose low power device only when on battery.
The low power device on laptops is the CPU and, especially when an external screen is attached, rendering creates significant load, which results in the fans spinning up. Using the discrete GPU when available alleviates the issue. Because it uses more power we only use the default device when not on battery.